% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crlapply-package.r
\docType{package}
\name{crlapply-package}
\alias{crlapply-package}
\title{crlapply}
\description{
Checkpoint/restart is a useful strategy for running expensive
functions in constrained environments (non-reliable hardware, restricted
time limits, etc). This package adds an lapply-like function that
automatically handles checkpointing.
}
\author{
Drew Schmidt
}
\keyword{Package}

