% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsapply.r
\name{crsapply}
\alias{crsapply}
\title{crsapply}
\usage{
crsapply(X, FUN, ..., checkpoint_file, checkpoint_freq = 1)
}
\arguments{
\item{X, FUN, ...}{Same as in \code{lapply()}.}

\item{checkpoint_file}{The checkpoint file.}

\item{checkpoint_freq}{The checkpoint frequency, i.e. how often the function will save checkpoints.
A frequency of 1 means that saves occur after every evaluation, 2 after
every other, 3 every third, and so on. Should a positive integer.}
}
\value{
A list.
}
\description{
An \code{sapply()}-like interface with automatic checkpoint/restart
functionality. A simple wrapper around \code{crlapply()}.
}
\details{
The checkpoint file is removed on successful completion.
}
\seealso{
\code{\link{crlapply}}
}
