% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crlapply.r
\name{crlapply}
\alias{crlapply}
\title{crlapply}
\usage{
crlapply(X, FUN, FILE, FREQ = 1, ...)
}
\arguments{
\item{X, FUN, ...}{Same as in \code{lapply()}.}

\item{FILE}{The checkpoint file.}

\item{FREQ}{The checkpoint frequency; a positive integer.}
}
\value{
A list.
}
\description{
An \code{lapply()}-like interface with automatic checkpoint/restart
functionality.
}
\details{
The checkpoint file is cleaned up on successful completion of
\code{crlapply()}
}

